--- extra/fsdp/utils.py
@@ -1,64 +1,0 @@
-from tinygrad import Tensor
-from typing import List, Callable, Optional
-from tinygrad.multi import MultiLazyBuffer
-
-def get_size(tensors: List[Tensor],
-             getter: Callable[[Tensor], int]=lambda t: t.nbytes(),
-             units: List[str]=["bytes", "KB", "MB", "GB"]):
-  size = sum([getter(t) if isinstance(t, Tensor) else t.size for t in tensors])
-  for i, unit in enumerate(units):
-    if size < 1000 or i == len(units) - 1: break
-    size /= 1000
-  return size, unit
-
-
-def print_lb(lb):
-  assert isinstance(lb, MultiLazyBuffer)
-  seen = set()
-  def repr_lb(lb, indent=0):
-    text = "\u001b[2m" if lb in seen else "\u001b[0m"
-    seen.add(lb)
-    def _indent(): return " " * indent
-    if isinstance(lb, MultiLazyBuffer):
-      text += _indent() + "MLB "
-      if lb.axis is not None:
-        text += _indent() + f"shape {lb.shape} axis {lb.axis}"
-        text += _indent() + f" bounds {lb.bounds}"
-      text += "\n"
-      indent += 4
-      text += _indent()
-      for i, _lb in enumerate(lb.lbs):
-        text += repr_lb(_lb, indent)
-        if i < len(lb.lbs) - 1: text += _indent()
-    else:     
-      text += "LB"
-      view = lb.st.views[0]
-      text += f" shape {view.shape}"
-      text += f" strides {view.strides}"
-      text += f" offset {view.offset}" if view.offset != 0 else ""
-      text += f" mask {view.mask}" if view.mask is not None else ""
-      text += "\n"
-      indent += 3
-      if hasattr(lb, "_base") and lb._base is not None:
-        text += _indent() + "_base: "
-        text += repr_lb(lb._base, indent + 7)
-        return text
-      else:
-        text += _indent()
-        text += f"{lb.device} "
-        text += f"\u001b[35m{lb.op}\u001b[30m "
-        text += f"arg: {lb.arg} " if lb.arg is not None else ""
-        text += f"size: {lb.size} \n"
-        if hasattr(lb, "buffer") and hasattr(lb.buffer, "_buf"):
-          text += _indent() + f"buf: {lb.buffer.nbytes} bytes\n"
-          return text
-        elif len(lb.srcs) > 0:
-          text += _indent() + "srcs: "
-          indent += 6
-          for i, src in enumerate(lb.srcs):
-            text += "\u001b[0m"
-            text += repr_lb(src, indent)
-            if i < len(lb.srcs) - 1: text += _indent()
-    return text
-  print(repr_lb(lb) + "\u001b[0m")
-
